+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   `TextField`: handle `aria-disabled=true` for an accessible disabled state
+-   `DatePickerField`: handle `aria-disabled=true` for an accessible disabled state
 
 ### Fixed